# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OleoControl",
        "description": "Documentación de la API de la aplicación OleoControl",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Listar todos los socios",
                "description": "Muestra todos los socios.\n\nEste método obtiene todos los socios de la base de datos y devuelve una respuesta JSON con un estado de éxito y los datos de los socios.",
                "operationId": "4e8230d39fe3529417c9a460cbd63a4c",
                "responses": {
                    "200": {
                        "description": "Lista de socios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MemberResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Miembros"
                ],
                "summary": "Crear un nuevo miembro",
                "description": "Crea un nuevo miembro junto con su usuario y rol correspondiente.\n\nEste método crea un nuevo miembro, su usuario asociado y le asigna el rol de \"Socio\". La creación se realiza dentro de una transacción para garantizar la integridad de los datos.\nLa respuesta incluye un estado de éxito y los datos del usuario creado.",
                "operationId": "87592872c1667344a26dc410cef5b4ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "dni",
                                    "email",
                                    "phone",
                                    "status"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678Z"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan.perez@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "612345678"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Miembro creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/members/{id}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Obtener los datos de un socio por ID",
                "description": "Muestra un socio específico por su ID.\n\nEste método recibe un ID de socio, busca el socio en la base de datos y devuelve una respuesta JSON con un estado de éxito y los datos del socio.",
                "operationId": "749194efa66230d3974f3419bc51c111",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del socio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Socio no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Miembros"
                ],
                "summary": "Actualizar datos de un socio",
                "description": "Actualiza un socio específico por su ID.\n\nEste método recibe una solicitud de actualización de socio, valida los datos y actualiza tanto el miembro como el usuario asociado en la base de datos.\nLa respuesta incluye un estado de éxito y los datos del socio actualizado.\nSi ocurre algún error, se realiza un rollback de la transacción y se devuelve un mensaje de error.",
                "operationId": "fca4ad568b04dd9ed139a41559860aed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "dni",
                                    "email",
                                    "phone",
                                    "status"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678Z"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan.perez@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "612345678"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Socio actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el socio"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Members"
                ],
                "summary": "Eliminar un socio",
                "description": "Elimina un socio específico por su ID.\n\nEste método recibe un ID de socio, busca el socio en la base de datos y elimina el socio.\nLa respuesta incluye un estado de éxito y un mensaje de éxito en formato JSON.",
                "operationId": "1fb9bc482ee522c0c7d14f5fb6e13ff6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Socio eliminado satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Socio no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MemberResource": {
                "title": "MemberResource",
                "description": "Recurso que representa la información de un miembro, incluyendo los datos de usuario asociados",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "member_number": {
                        "type": "string",
                        "example": "12345"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "username": {
                                "type": "string",
                                "example": "johndoe"
                            },
                            "first_name": {
                                "type": "string",
                                "example": "John"
                            },
                            "last_name": {
                                "type": "string",
                                "example": "Doe"
                            },
                            "dni": {
                                "type": "string",
                                "example": "12345678X"
                            },
                            "email": {
                                "type": "string",
                                "example": "johndoe@example.com"
                            },
                            "phone": {
                                "type": "string",
                                "example": "123456789"
                            },
                            "status": {
                                "type": "string",
                                "example": "active"
                            },
                            "roles": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Socio"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "UserResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "juan.perez.z"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "dni": {
                        "type": "string",
                        "example": "12345678Z"
                    },
                    "email": {
                        "type": "string",
                        "example": "juan.perez@ejemplo.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "123456789"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Socio"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "member": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "member_number": {
                                "type": "integer",
                                "example": 1001
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "employee": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "department": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Contabilidad"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación Bearer (usado por Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}